Working screen space clipping but there are some horrible glitches

+++ R/0D/libs/sdw/CanvasPoint.cpp
@@ -15,7 +15,7 @@ CanvasPoint::CanvasPoint(float xPos, float yPos)
   texturePoint = TexturePoint(-1,-1);
 }
 
-CanvasPoint::CanvasPoint(float xPos, float yPos, float pointDepth)
+CanvasPoint::CanvasPoint(float xPos, float yPos, double pointDepth)
 {
   x = xPos;
   y = yPos;
